Change consul puppet module source to include patch
The puppet consule package has a class to reload the
configuration, but it can fail if the reload happens
as consul is booting. I have written a patch that
introduces sleep between each reload try to give the
time to consul to finish booting. At the time of
writting this commit, the patch has yet been merged
in the original repo and therefore not released to
puppetforge. I am pointing the source of puppet
consul to my own fork until the patch is merged.