Configure resolv.conf with NetworkManager
When the search line was missing from resolv.conf,
puppet file_line was appending the line at the end
of the file, preventing dns resolving to work
correctly. We will let NetworkManager do its thing
to configure resolv.conf and simply add a file to
instruct the search domain and the new nameservers
to NetworkManager.