Use nodes instead of instances to config slurm
The $instances variable contains all instances that
are part of the cluster, but only instances with a
'node' tag are compute nodes that needs to be considered
when configuring Slurm. Using $instances instead of $nodes
was functional but not optimal and create weird quirk like
compute node weights in nodes.conf not being defined on a
continuous range because an instance configuration would be
considered when computing weights, but would not show up in
nodes.conf because it was not associated with a compute node.