Disable Slurm gids cache
This is to avoid issues when adding multiple groups to a user
with FreeIPA web interface where the groups will then be missing
during a job but not when sshing on the node.

The Slurm gid cache prevents overloading the LDAP server, but given
most MC cluster scale, that should not be an issue for now.