Increase power save interval from default 10 to 30 seconds
This avoids fragmenting the Terraform Cloud requests in too many.
When for example requesting the power up of 5 nodes in a single
command, with the default interval of 10, Slurm will launch two
terraform runs, while with an interval of 30, a single run will
be created, reducing the overall latency of the nodes being ready
to run jobs.